Resolved the rest of the non move-constructor related issues

+++ Plugins/RPGMenus/Source/RPGMenus/Private/Windows/MessageWindow.cpp
@@ -135,7 +135,7 @@ void UMessageWindow::ResizeWindow() {
     if (SizeBox != nullptr && DisplayTextWidget != nullptr) {
         auto TextHeight = static_cast<float>(DisplayTextWidget->GetTextSize("Sample").Y);
         auto DisplayTextPadding = DisplayTextWidget->GetDisplayTextPadding();
-        SizeBox->SetHeightOverride(TextHeight * LinesToShow + DisplayTextPadding.Top + DisplayTextPadding.Bottom);
+        SizeBox->SetHeightOverride(TextHeight * static_cast<float>(LinesToShow) + DisplayTextPadding.Top + DisplayTextPadding.Bottom);
     }
 }
